% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bpplot}
\alias{bpplot}
\title{Plot Daily Blood-Pressure Readings}
\usage{
bpplot(
  files,
  legend = TRUE,
  smooth = nrow(BP) >= 15,
  span = seq(0.1, 0.9, by = 0.1),
  confint = smooth,
  level = c(0.5, 0.9),
  alpha = 0.25,
  vlines = NULL,
  xlab = "Date",
  ylab = "Blood Pressure (mm Hg)",
  show.main = TRUE
)
}
\arguments{
\item{files}{character vector of (paths to) BP files.}

\item{legend}{(logical) if \code{TRUE} (the default) show a legend; if \code{FALSE} label
systolic and diastolic values directly.}

\item{smooth}{(logical) show loess smooths (default \code{TRUE} if there are
at least 15 dates)?}

\item{span}{for loess smooths; if there is more than one, the span is
selected by 10-fold cross-validation; the default is 9 values
between 0.1 and 0.9. Cross-validation doesn't work well if there
are discontinuities.}

\item{confint}{(logical) show pointwise confidence envelope(s) around smooths?
the default is the value of \code{smooth}.}

\item{level}{for confidence envelope(s); the default is 0.5 and 0.9.}

\item{alpha}{transparency for confidence envelopes (cumulative); the default
is 0.25.}

\item{vlines}{list of lists for vertical lines to draw
(the \code{"at"} component of each sub-list), labels for
the lines (the \code{"text"} component), and (optionally, the
\code{"at"} component) for vertical adjustment of text;
see examples.}

\item{xlab, ylab}{horizontal axis and vertical axis labels.}

\item{show.main}{display main label (logical, default \code{TRUE}).}
}
\value{
invisibly returns a data frame with columns for date, systolic
blood pressure, and diastolic blood pressure.
If cross-validation is used to pick the span, a table of spans
and CV mean-squared error values is printed.
}
\description{
\code{bpplot()} graphs daily blood-pressure readings. For more detailed
analysis of blood-pressure data, see the \pkg{bp} package
(\url{https://github.com/johnschwenck/bp} and \url{https://cran.r-project.org/package=bp}).
}
\examples{
dir <- system.file("etc", package="bpplot")
(files <- list.files(dir)) # example data files provided
files <- paste0(dir, "/", files)
cat(paste(readLines(files[5], 10), "\n")) # part of one data file

bpplot(files[3:5],
       span=0.3,
       vlines=list(list(at="2025-07-08",
                        text="pause \u2192\nDrug      "),
                   list(at="2025-07-15",
                        text="resume\u2192\nDrug      ",
                        adjust=15))
)

bpplot(files[1:3], legend=FALSE)
}
